+++ src/settings/MappingProfiles/detailsSections/edit/tests/MappingInvoiceDetails.test.js
@@ -1,9 +1,11 @@
 import React from 'react';
 import { fireEvent } from '@testing-library/react';
 
-import { renderWithIntl } from '@folio/stripes-data-transfer-components/test/jest/helpers';
-
 import '../../../../../../test/jest/__mock__';
+
+import { renderWithIntl } from '@folio/stripes-data-transfer-components/test/jest/helpers';
+import { FOLIO_RECORD_TYPES } from '@folio/stripes-data-transfer-components';
+
 import {
   renderWithReduxForm,
   translationsProperties,
@@ -14,7 +16,6 @@ import {
   getInitialFields,
   getInitialDetails,
 } from '../../../initialDetails';
-import { FOLIO_RECORD_TYPES } from '../../../../../components';
 
 jest.mock('.../../../../../components/FieldOrganization/FieldOrganization', () => ({ onSelect }) => (
   <div>