+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * Add required info icons to Holdings field mapping profile (UIDATIMP-1292)
 * Add required info icons to Item field mapping profile (UIDATIMP-1293)
 * Allow edit and duplicate Default - Create SRS MARC Authority job profile (UIDATIMP-1318)
+* Change imports for ListTemplate component in ui-data-import (UIDATIMP-1301)
 * Improve mapping profile for Orders (UIDATIMP-1333)
 * Order field mapping profile: Include default value for PO line limit setting (UIDATIMP-1337)